+++ src/clients/cs_deleteFileInfoListByFileIdList.cpp
@@ -1,5 +1,6 @@
 #include "contentServer/corba/client/ClientImplementation.h"
 #include "contentServer/http/client/ClientImplementation.h"
+#include "contentServer/redis/RedisImplementation.h"
 #include "grid-files/common/Exception.h"
 #include "grid-files/common/GeneralFunctions.h"
 
@@ -12,7 +13,7 @@ int main(int argc, char *argv[])
   {
     if (argc < 3)
     {
-      fprintf(stdout,"USAGE: cs_deleteFileInfoListByFileIdList <sessionId> <fileId1> [<fileId2> .. <fileIdN>] [-http <url>]\n");
+      fprintf(stdout,"USAGE: cs_deleteFileInfoListByFileIdList <sessionId> <fileId1> [<fileId2> .. <fileIdN>] [[-http <url>]|[-redis <address> <port> <tablePrefix>]]\n");
       return -1;
     }
 
@@ -30,6 +31,16 @@ int main(int argc, char *argv[])
     {
       ContentServer::HTTP::ClientImplementation service;
       service.init(argv[argc-1]);
+
+      startTime = getTime();
+      result = service.deleteFileInfoListByFileIdList(sessionId,fileIdList);
+      endTime = getTime();
+    }
+    else
+    if (argc > 4  &&  strcmp(argv[argc-4],"-redis") == 0)
+    {
+      ContentServer::RedisImplementation service;
+      service.init(argv[argc-3],atoi(argv[argc-2]),argv[argc-1]);
 
       startTime = getTime();
       result = service.deleteFileInfoListByFileIdList(sessionId,fileIdList);
